tests/wpt/web-platform-tests/css/css-transitions/parsing/transition-delay-invalid.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Transitions: parsing transition-delay with invalid values</title>
<link rel="help" href="https://drafts.csswg.org/css-transitions/#propdef-transition-delay">
<meta name="assert" content="transition-delay supports only the grammar '<time> #'.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
</head>
<body>
<script>
test_invalid_value("transition-delay", 'infinite');
test_invalid_value("transition-delay", '0');
test_invalid_value("transition-delay", '500ms 0.5s');
</script>
</body>
</html>
